requirements.txt
asgiref==3.8.1
attrs==25.3.0
coverage==7.8.2
Django==5.2.1
django-debug-toolbar==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
flake8==7.2.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
